+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [1.8.0] - 2026-01-13
 
 ### Added
+- Event-driven detection of new/updated labeled Kubernetes secrets. (CNJR-12568)
 - Support for onboarding labeled Kubernetes secrets. (CNJR-12567)
 
 ## [1.7.8] - 2025-12-23